LPD-48502 Move normalizeName to objectDefinitionUtils in object-js-components-web
