+++ modules/apps/object/object-js-components-web/src/main/resources/META-INF/resources/index.ts
@@ -61,3 +61,4 @@ export * as arrayUtils from './utils/array';
 export * as constantsUtils from './utils/constants';
 export * as datetimeUtils from './utils/datetime';
 export * as errorsUtils from './utils/errors';
+export * as objectDefinitionUtils from './utils/objectDefinition';
