+++ modules/apps/object/object-js-components-web/src/main/resources/META-INF/resources/utils/objectDefinition.ts
@@ -1,0 +1,17 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+import {firstLetterUppercase, removeAllSpecialCharacters} from './string';
+
+export function normalizeName(str: string) {
+	const split = str.split(' ');
+
+	const capitalizeFirstLetters = split.map((str: string) =>
+		firstLetterUppercase(str)
+	);
+	const join = capitalizeFirstLetters.join('');
+
+	return removeAllSpecialCharacters(join);
+}
